fix(repl): fix validate-gate, preserve description on overwrite, extract save/undo
- Validate-gate now compares in-memory error count against on-disk baseline,
  allowing saves when the model was already invalid before the REPL session
- Empty description input on overwrite now keeps the existing description
  instead of silently clearing it
- saveCommand, patchSave, elementsEqual, undoCommand, saveUndo extracted to
  repl_save.go to keep repl.go within the 500-line project convention
- Tests for new behaviors in repl_save_test.go; old duplicates removed from
  repl_test.go

Closes #410
